// src/components/dose-timeline/dose-timeline-types.ts
export interface PhaseTimings {
  onsetEnd: number;
  comeupEnd: number;
  peakEnd: number;
  offsetEnd: number;
  afterglowDuration: number;
  afterglowEnd: number;
  totalDuration: number;
}